+++ src/Formie.php
@@ -99,7 +99,7 @@ class Formie extends Plugin
 
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '3.4.5';
+    public string $schemaVersion = '3.4.6';
     public string $minVersionRequired = '2.1.5';