+++ src/base/Field.php
@@ -765,6 +765,10 @@ abstract class Field extends SavableComponent implements CraftFieldInterface, Fi
 
         $this->trigger(static::EVENT_MODIFY_DEFAULT_VALUE, $event);
 
+        if (is_string($event->value)) {
+            $event->value = trim($event->value);
+        }
+
         return $event->value;
     }