+++ src/integrations/emailmarketing/Klaviyo.php
@@ -26,6 +26,7 @@ class Klaviyo extends EmailMarketing
         return Craft::t('formie', 'Klaviyo');
     }
 
+
     // Properties
     // =========================================================================
 
@@ -46,60 +47,79 @@ class Klaviyo extends EmailMarketing
         $settings = [];
 
         try {
-            $lists = $this->request('GET', 'v2/lists');
+            $response = $this->request('GET', 'lists');
+            $lists = $response['data'] ?? [];
 
             foreach ($lists as $list) {
                 $listFields = [
                     new IntegrationField([
-                        'handle' => '$first_name',
+                        'handle' => 'first_name',
                         'name' => Craft::t('formie', 'First Name'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$last_name',
+                        'handle' => 'last_name',
                         'name' => Craft::t('formie', 'Last Name'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$email',
+                        'handle' => 'email',
                         'name' => Craft::t('formie', 'Email'),
                         'required' => true,
                     ]),
                     new IntegrationField([
-                        'handle' => '$phone_number',
+                        'handle' => 'phone_number',
                         'name' => Craft::t('formie', 'Phone Number'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$city',
+                        'handle' => 'city',
                         'name' => Craft::t('formie', 'City'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$region',
+                        'handle' => 'region',
                         'name' => Craft::t('formie', 'Region'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$country',
+                        'handle' => 'country',
                         'name' => Craft::t('formie', 'Country'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$zip',
+                        'handle' => 'zip',
                         'name' => Craft::t('formie', 'Zip'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$organization',
+                        'handle' => 'organization',
                         'name' => Craft::t('formie', 'Organization'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$title',
+                        'handle' => 'title',
                         'name' => Craft::t('formie', 'Title'),
                     ]),
                     new IntegrationField([
-                        'handle' => '$source',
+                        'handle' => 'source',
                         'name' => Craft::t('formie', 'Source'),
+                    ]),
+                    new IntegrationField([
+                        'handle' => 'sms_consent',
+                        'name' => Craft::t('formie', 'Consent To Send SMS'),
+                        'type' => IntegrationField::TYPE_BOOLEAN,
+                        'options' => [
+                            'label' => Craft::t('formie', 'Consent To Send SMS'),
+                            'options' => [
+                                [
+                                    'label' =>  Craft::t('formie', 'Yes'),
+                                    'value' => true,
+                                ],
+                                [
+                                    'label' =>  Craft::t('formie', 'No'),
+                                    'value' => false,
+                                ],
+                            ],
+                        ],
                     ]),
                 ];
 
                 $settings['lists'][] = new IntegrationCollection([
-                    'id' => $list['list_id'],
-                    'name' => $list['list_name'],
+                    'id' => $list['id'],
+                    'name' => $list['attributes']['name'],
                     'fields' => $listFields,
                 ]);
             }
@@ -115,33 +135,68 @@ class Klaviyo extends EmailMarketing
         try {
             $fieldValues = $this->getFieldMappingValues($submission, $this->fieldMapping);
 
-            // Create the profile first, with the Public API
+            // Create or update a Profile first
             $payload = [
-                'token' => App::parseEnv($this->publicApiKey),
-                'properties' => $fieldValues,
+                'data' => [
+                    'type' => 'profile',
+                    'attributes' => $fieldValues,
+                ],
             ];
 
-            $response = $this->deliverPayload($submission, 'identify', $payload);
+            $response = $this->deliverPayload($submission, 'profile-import', $payload);
 
             if ($response === false) {
                 return true;
             }
 
-            // Subscribe the user to the list
-            $email = ArrayHelper::remove($fieldValues, '$email');
-            $source = ArrayHelper::remove($fieldValues, '$source');
+            $profileId = $response['data']['id'] ?? '';
 
-            $profile = ['email' => $email];
+            if (!$profileId) {
+                Integration::error($this, Craft::t('formie', 'Missing return “profileId” {response}. Sent payload {payload}', [
+                    'response' => Json::encode($response),
+                    'payload' => Json::encode($payload),
+                ]), true);
 
-            if ($source) {
-                $profile['$source'] = $source;
+                return false;
             }
 
+            // Extract any consent settings
+            $smsConsent = ArrayHelper::remove($fieldValues, 'sms_consent');
+
+            $profile = $fieldValues;
+            $profile['subscriptions']['email']['marketing']['consent'] = 'SUBSCRIBED';
+
+            if ($smsConsent) {
+                $profile['subscriptions']['sms']['marketing']['consent'] = 'SUBSCRIBED';
+            }
+
+            // Subscribe the user to the list
             $payload = [
-                'profiles' => [$profile],
+                'data' => [
+                    'type' => 'profile-subscription-bulk-create-job',
+                    'attributes' => [
+                        'profiles' => [
+                            'data' => [
+                                [
+                                    'type' => 'profile',
+                                    'id' => $profileId,
+                                    'attributes' => $profile,
+                                ],
+                            ],
+                        ],
+                    ],
+                    'relationships' => [
+                        'list' => [
+                            'data' => [
+                                'type' => 'list',
+                                'id' => $this->listId,
+                            ],
+                        ],
+                    ],
+                ],
             ];
 
-            $response = $this->deliverPayload($submission, "v2/list/{$this->listId}/subscribe", $payload);
+            $response = $this->deliverPayload($submission, 'profile-subscription-bulk-create-jobs', $payload);
 
             if ($response === false) {
                 return true;
@@ -158,7 +213,7 @@ class Klaviyo extends EmailMarketing
     public function fetchConnection(): bool
     {
         try {
-            $response = $this->request('GET', 'v2/lists');
+            $response = $this->request('GET', 'lists');
         } catch (Throwable $e) {
             Integration::apiError($this, $e);
 
@@ -176,8 +231,9 @@ class Klaviyo extends EmailMarketing
 
         return $this->_client = Craft::createGuzzleClient([
             'base_uri' => 'https://a.klaviyo.com/api/',
-            'query' => [
-                'api_key' => App::parseEnv($this->privateApiKey),
+            'headers' => [
+                'Authorization' => 'Klaviyo-API-Key ' . App::parseEnv($this->privateApiKey),
+                'revision' => '2024-05-15',
             ],
         ]);
     }